🚚rename : Policy의 이용약관 상수 파일 위치 변경
이요약관 상수가 PolicyPage, RegisterPag의 PolicyModal에서 쓰이므로 공통 파일인 constants 하위 폴더로 위치를 변경 후 import 경로를 수정했습니다.
